requirements.txt
pytest
pylint
jupyter
pytest-cov
pandas
nbval
click
flask
requests
numpy
matplotlib
tensorflow
seaborn
scikit-learn
earthengine-api
opencv-contrib-python-headless
gunicorn
google-cloud-bigquery
